backend/requirements/base.txt
# Django and DRF
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0

# Database
psycopg[binary]>=3.1,<4.0

# Configuration
python-decouple>=3.8,<4.0
dj-database-url>=2.1,<3.0

# Data/ML
pandas>=2.1,<3.0
numpy>=1.26,<2.0
scikit-learn>=1.4,<2.0
openpyxl>=3.1,<4.0
joblib>=1.3,<2.0
shap>=0.44,<1.0

# PDF Report Generation
weasyprint>=67.0,<68.0
matplotlib>=3.8,<4.0

# Gemini API
google-generativeai>=0.3,<1.0

# File handling
python-magic>=0.4,<1.0

# Celery & Redis
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0
